# sovd-server/openapi.yaml
openapi: 3.1.0
info:
  title: sovd-server
  description: HTTP/REST SOVD server for the Eclipse OpenSOVD core stack
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
...
  /sovd/v1/components/{component_id}/faults:
    get:
      tags:
      - fault-handling
      summary: '`GET /sovd/v1/components/{component_id}/faults` — list faults.'
      description: |-
        # Errors

        Returns 404 if the component is unknown; other
        [`SovdError`](sovd_interfaces::SovdError) values are mapped via
        [`ApiError`].
      operationId: getFaults
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: severity
        in: query
        description: Severity upper bound (exclusive)
        required: false
        schema:
          type: integer
          format: int32
      - name: scope
        in: query
        description: Scope string
        required: false
        schema:
          type: string
      - name: status_key
        in: query
        description: Status filter in key:value form
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: 1-based page number (default 1)
        required: false
        schema:
          type: string
      - name: page-size
        in: query
        description: Page size (default 50)
        required: false
        schema:
          type: string
      responses:
        '200':
          description: List of faults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOfFaults'
        '400':
          description: Invalid pagination request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '404':
          description: Component not found
    delete:
      tags:
      - fault-handling
      summary: '`DELETE /sovd/v1/components/{component_id}/faults` — clear every fault.'
      description: |-
        # Errors

        Returns 404 if the component is unknown.
      operationId: deleteAllFaults
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: All faults cleared
        '404':
          description: Component not found
  /sovd/v1/components/{component_id}/faults/{fault_code}:
    get:
      tags:
      - fault-handling
      summary: |-
        `GET /sovd/v1/components/{component_id}/faults/{fault_code}` — fault
        details for one code.
      description: |-
        # Errors

        Returns 404 if the component or fault code is unknown.
      operationId: getFaultById
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: fault_code
        in: path
        description: Native fault code (string)
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Fault details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaultDetails'
        '404':
          description: Fault not found
    delete:
      tags:
      - fault-handling
      summary: |-
        `DELETE /sovd/v1/components/{component_id}/faults/{fault_code}` — clear
        one fault.
      description: |-
        # Errors

        Returns 404 if the component or fault code is unknown.
      operationId: deleteFaultById
      parameters:
      - name: component_id
        in: path
        description: Stable component identifier
        required: true
        schema:
          type: string
      - name: fault_code
        in: path
        description: Native fault code (string)
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Fault cleared
        '404':
          description: Fault not found
...
components:
  schemas:
...
    DataError:
      type: object
      description: |-
        SOVD partial-error envelope, used when only one element of a structured
        response failed.

        Provenance: `commons/errors.yaml#DataError`.

        `path` is a JSON Pointer pointing at the offending field of the response
        body. `error` describes the failure in more detail.
      required:
      - path
      properties:
        error:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/GenericError'
            description: Detailed error describing why `path` failed.
        path:
          type: string
          description: JSON Pointer to the offending element of the response.
...
    Fault:
      type: object
      description: |-
        A single SOVD fault entry.

        Provenance: `faults/types.yaml#Fault`.

        `code` and `fault_name` are required by the spec. Every other field is
        optional and may be omitted by the entity.
      required:
      - code
      - fault_name
      properties:
        code:
          type: string
          description: Native fault code, e.g. `"0012E3"`, `"P102"`, `"modelMissing"`.
        display_code:
          type:
          - string
          - 'null'
          description: Display representation of `code`, e.g. `"P102"`.
        fault_name:
          type: string
          description: Human-readable fault name.
        fault_translation_id:
          type:
          - string
          - 'null'
          description: Identifier for translating `fault_name`.
        scope:
          type:
          - string
          - 'null'
          description: Scope of the fault (e.g. user-defined fault memory). Capability-defined.
        severity:
          type:
          - integer
          - 'null'
          format: int32
          description: 'Severity. Spec convention: `1=FATAL, 2=ERROR, 3=WARN, 4=INFO`.'
        status:
          description: |-
            OEM-specific status key/value object. For classic ECUs typically
            includes the UDS DTC status-byte bits.
        symptom:
          type:
          - string
          - 'null'
          description: OEM-specific symptom / failure-mode description.
        symptom_translation_id:
          type:
          - string
          - 'null'
          description: Identifier for translating `symptom`.
        tags:
          type:
          - array
          - 'null'
          items:
            type: string
          description: Tags attached to this fault entry.
    FaultDetails:
      type: object
      description: |-
        Response body for `GET /{entity-collection}/{entity-id}/faults/{fault-code}`.

        Provenance: `faults/responses.yaml#FaultDetails` (inline schema — the
        spec does not give it a name in `components/schemas`, so we coin one).
      required:
      - item
      properties:
        environment_data:
          description: OEM-specific environment / freeze-frame data.
        errors:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/DataError'
          description: Errors related to `environment_data`, if any.
        extras:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ResponseExtras'
            description: |-
              Soft-fail marker per ADR-0018 rule 4. Absent on the nominal
              path; set with `stale: true` when a degraded response is
              served. Extra (per ADR-0006): not part of ISO 17978-3.
        item:
          $ref: '#/components/schemas/Fault'
          description: The fault entry itself.
        schema:
          description: Optional embedded JSON Schema describing the response shape.
...
    GenericError:
      type: object
      description: |-
        SOVD generic error envelope.

        Provenance: `commons/errors.yaml#GenericError`.

        Returned in the body of any non-2xx SOVD response. The `error_code` is
        a SOVD-standardised code; `vendor_code` is filled when the error is
        vendor-specific. `parameters` carries open key/value detail.
      required:
      - error_code
      - message
      properties:
        error_code:
          type: string
          description: SOVD standardised error code.
        message:
          type: string
          description: Human-readable problem description.
        parameters:
          description: Additional key/value detail about the error.
        translation_id:
          type:
          - string
          - 'null'
          description: Identifier for translating `message`.
        vendor_code:
          type:
          - string
          - 'null'
          description: Vendor-specific error code, set when `error_code` is vendor-specific.
...
    ListOfFaults:
      type: object
      description: |-
        Response body for `GET /{entity-collection}/{entity-id}/faults`.

        Provenance: `faults/types.yaml#ListOfFaults`.

        `schema` is only present if the request was made with
        `?include-schema=true`; we carry it as an opaque JSON value to avoid
        modelling the full `OpenAPI` 3.1 meta-schema.
      required:
      - items
      properties:
        extras:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ResponseExtras'
            description: |-
              Soft-fail marker per ADR-0018 rule 4. Absent on the nominal
              path so the spec-pure shape is preserved; set to
              `Some(ResponseExtras { stale: true, .. })` when the backend
              returned a last-known snapshot because a fresh query failed
              or a retry budget was exhausted.

              Extra (per ADR-0006): not part of ISO 17978-3.
        items:
          type: array
          items:
            $ref: '#/components/schemas/Fault'
          description: All faults that match the request filter.
        next_page:
          type:
          - integer
          - 'null'
          format: int32
          description: Cursor for the next page, expressed as the next 1-based page number.
          minimum: 0
        schema:
          description: Optional embedded JSON Schema describing the response shape.
        total:
          type:
          - integer
          - 'null'
          format: int64
          description: Total number of faults that matched the filter before paging.
          minimum: 0
...
    ResponseExtras:
      type: object
      description: |-
        Soft-fail metadata attached to a SOVD response body per ADR-0018.

        Extra (per ADR-0006): not part of ISO 17978-3. Carried as an
        optional, flatten-friendly struct under the `extras` field of
        select response types.

        # Field conventions

        - `stale` is `true` when the backend returned something other
          than a fresh, authoritative result (e.g. the last-known cache
          snapshot, a degraded retry result, or an aggregated fan-out
          with one host missing).
        - `age_ms` is set when the response came from a cache; it is the
          age of the snapshot in milliseconds.
        - `host_unreachable` is a sorted list of component ids whose
          remote host could not be reached during a `sovd-gateway`
          fan-out. An empty list or `None` means "no known partial
          failures."
      properties:
        age_ms:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Age of the cached snapshot in milliseconds, when `stale` was
            set because a cache was served instead of fresh data.
          minimum: 0
        host_unreachable:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/String'
          description: |-
            Component ids whose backing remote host was unreachable during
            this request. Used by `sovd-gateway::list_components` fan-out
            per ADR-0018 rule 5.

            Wire type is `Vec<String>` (not `Vec<ComponentId>`) because
            `ComponentId` does not derive `utoipa::ToSchema`; call sites
            that need the typed form convert via
            [`ResponseExtras::host_unreachable_typed`].
        stale:
          type: boolean
          description: |-
            `true` when this response is a degraded / last-known fallback
            rather than a full-strength answer.
...
    String:
      type: string
...